Split websocket client into separate module
